frontend/electrum: convert to TypeScript and FunctionalComponent
